+++ examples/core/audit/otel.py
@@ -1,8 +1,8 @@
 """
-Ragbits Core Example: OpenTelemetry
+Ragbits Core Example: OpenTelemetry Audit
 
-This example demonstrates how to collect traces and metrics using Ragbits audit module.
-We run the LLM generation several times to collect telemetry data, and then export it to the OpenTelementry collector and visualize it in Grafana.
+This example demonstrates how to collect traces and metrics using Ragbits audit module with OpenTelemetry.
+We run the LLM generation several times to collect telemetry data, and then export it to the OpenTelemetry collector and visualize it in Grafana.
 
 The script exports traces to the local OTLP collector running on http://localhost:4317.
 The recommended way to run it is using the official Docker image:
@@ -33,9 +33,11 @@ To visualize the metrics collected by Ragbits, follow these steps:
 #     "ragbits-core[otel]",
 #     "opentelemetry-sdk",
 #     "opentelemetry-exporter-otlp-proto-grpc",
+#     "google-auth>=2.35.0",
 #     "tqdm",
 # ]
 # ///
+
 import asyncio
 from collections.abc import AsyncGenerator