+++ packages/ragbits-core/src/ragbits/core/audit/metrics/__init__.py
@@ -35,13 +35,21 @@ def set_metric_handlers(handlers: Handler | list[Handler]) -> None:
         if isinstance(handler, MetricHandler):
             _metric_handlers.append(handler)
         elif isinstance(handler, str):
-            if handler == "otel":
-                from ragbits.core.audit.metrics.otel import OtelMetricHandler
+            match handler.lower():
+                case "otel":
+                    from ragbits.core.audit.metrics.otel import OtelMetricHandler
 
-                if not any(isinstance(item, OtelMetricHandler) for item in _metric_handlers):
-                    _metric_handlers.append(OtelMetricHandler())
-            else:
-                raise ValueError(f"Not found handler: {handler}")
+                    if not any(isinstance(item, OtelMetricHandler) for item in _metric_handlers):
+                        _metric_handlers.append(OtelMetricHandler())
+
+                case "logfire":
+                    from ragbits.core.audit.metrics.logfire import LogfireMetricHandler
+
+                    if not any(isinstance(item, LogfireMetricHandler) for item in _metric_handlers):
+                        _metric_handlers.append(LogfireMetricHandler())
+
+                case _:
+                    raise ValueError(f"Not found handler: {handler}")
         else:
             raise TypeError(f"Invalid handler type: {type(handler)}")