+++ typescript/ui/src/plugins/FeedbackPlugin/index.tsx
@@ -1,0 +1,10 @@
+import { lazy } from "react";
+import { createPlugin } from "../../core/utils/plugins/utils";
+
+export const FeedbackFormPluginName = "FeedbackFormPlugin";
+export const FeedbackFormPlugin = createPlugin({
+  name: FeedbackFormPluginName,
+  components: {
+    FeedbackForm: lazy(() => import("./FeedbackForm")),
+  },
+});